Make upsert metadata manager pluggable
