+++ pinot-segment-local/src/main/java/org/apache/pinot/segment/local/upsert/RecordLocation.java
@@ -24,6 +24,7 @@ import org.apache.pinot.segment.spi.IndexSegment;
 /**
  * Indicate a record's location on the local host.
  */
+@SuppressWarnings("rawtypes")
 public class RecordLocation {
   private final IndexSegment _segment;
   private final int _docId;
